Fix #12 - add 1 func : split_morse() - assemble lght_up and show_morse_code to light or sound
