+++ raspberry.py
@@ -1,31 +1,31 @@
-import time
-from machine import Pin
 import api
+import time
+import RPi.GPIO as GPIO
 
 
-class ESP32:
+class Raspberry:
     def __init__(self):
         self.morse_time = api.MorseTime()
-        self.button = Pin(17, Pin.IN, Pin.PULL_UP)
-        self.led = Pin(16, Pin.OUT)
+        GPIO.setmode(GPIO.BCM)
+        GPIO.setup(5, GPIO.OUT, initial=GPIO.LOW)
 
     def light(self, value: str):
         splited = api.text_to_morse(value)
         for word in splited:
             for symbol in word:
                 if symbol == ".":
-                    self.led.value(1)
+                    GPIO.output(5, GPIO.HIGH)
                     time.sleep(self.morse_time.court)
-                    self.led.value(0)
+                    GPIO.output(5, GPIO.LOW)
                     time.sleep(self.morse_time.very_short)
                 elif symbol == "-":
-                    self.led.value(1)
+                    GPIO.output(5, GPIO.HIGH)
                     time.sleep(self.morse_time.long)
-                    self.led.value(0)
+                    GPIO.output(5, GPIO.LOW)
                     time.sleep(self.morse_time.very_short)
             time.sleep(self.morse_time.court)
 
 
 if __name__ == "__main__":
-    traducteur_type = ESP32()
-    traducteur_type.light("TEST")
+    traducteur_type = Raspberry()
+    traducteur_type.light(input("quel est le texte que voulez vous afficher ?"))
